Add ARMED run state bwtween IDLE and RECORD
When a START command is received in the IDLE state, the state now
transitions to the ARMED state.  In the ARMED state packets are dropped
until getting a "startable" packet.  Once a startable packet is
received, the state transitions to RECORD.  See the comments in the diff
for more details.