Functioning multi-beam net thread
Multi-beam (mb1) data are saved to raw file with channel dimension
changing faster than time dimension.  This is different from the
canonical RAW file format, so a new header variable is defined (CHANMAJ).
When CHANMAJ is 0 (or undefined), the RAW file data are ordered normally
with the time dimension changing faster than the channel dimension.
When CHANMAJ is 1, the RAW file data are ordered with channel dimension
changing faster than the time dimension.