Ignore FIDs that exceed NANTS
When the number of antennas in a MeerKAT subarray is not a power of two,
the number of MeerKAT F Engines that participate is actually the next
power of two.  For example, a 6 antenna subarray would have 8 F Engines
sending packets.  The extra F engines are sending copies of the data
from the last antenna (presumably to simplify the FPGA based X engines).
Packets from these extra F engines must be ignored since they will
appear to belong in areas of memory that exceed expectations (i.e. lead
to segfaults).