Ensure OBSNCHAN is (re-)calculated at proper times
MeerKAT OBSNCHAN depends on NANTS, NSTRM, and HNCHAN, which are part of
the "obs_info" structure and come from the coordinator via the status
buffers.  OBSNCHAN should be recalculated whenever these values are read
from the status buffer (and are valid).  In theory, it only needs to be
recalculated if any of the factors change, but testing for that is
probably as time consuming as just recomputing every time.