Add another throughput metric (PKTBLKMS)
PKTBLKMS is a moving average of the number of milliseconds that the
spead thread takes between getting a filled block from the pkt thread
and freeing the block back to the pkt thread.  The moving average is
computed over N_INPUT_BLOCKS blocks and updated in the status buffer
every time the input block index wraps around.