Add hpguppi_ibverbs_pkt_thread.{c,h}
This is a very generic IB Verbs based packet reception thread.  This may
be worth moving into Hashpipe proper as a generic network input thread,
but that would require a generic input databuffer for this purpose (i.e.
not an hpguppi_input_databuf).

Still need to add functions for creating/destroying flows.