PKSUWL now requires DESTIP to be set at start time
This was an attempt to minimize problems when status buffer fields were
getting clobbered by non-thread-safe hgetXX() calls.  The hgetXX() calls
have been made thread-safe so this change is no longer helpful in that
regard and could be reverted, but given the static configuration of the
PKSUWL receiver, there's not much benefit to having DESTIP be
dynamically configurable.