Exit on error when writing to output file
Writing to the output file when the filesystem is full eventually leads
to the pipeline hanging (for reasons that are not fully understood).
When this happens, the hung process will not respond to changes in
DESTIP that indicate that multicast groups should be unsubscribed, which
in turn can lead to "phantom" subscriptions that can lead to an
oversubscribed state in the switching fabric which can lead to packet
loss.