Add more hashpipe libs to hpguppi_daq_la_LIBADD
This links some additional hashpipe related libraries to the
`hpguppi_daq.so` plug-in.  These libraries are already loaded when
hashpipe loads `hpguppi_daq.so` as a plug-in, so symbols from these
libraries are found in that case.  When loading `hpguppi_daq.so` outside
of hashpipe (e.g. via `dlopen()`) these libraries are not already loaded
so symbols from them cannot be found unless these libraries are linked
into `hgpuppi_daq.so`.  Adding these libraries to
`hpguppi_daq_la_LIBADD` achieves that goal.